Web/requirements.txt
django
djangorestframework
pyrebase